netaxe_ipam/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
csscompressor==0.9.5
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1.2
django-allauth==0.51.0
django-auth-ldap==4.1.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cidrfield==0.2.0
django-compress-staticfiles==1.0.1b0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-formtools==2.4
django-guardian==2.1.0
django-import-export==2.9.0
django-model-utils==4.2.0
django-organizations==2.0.2
django-phonenumber-field==6.0.0
django-redis==5.2.0
django-reversion==5.0.3
django-sesame==2.4
django-simpleui==2022.7.29
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
drf-yasg==1.20.3
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.12.3
idna==3.4
importlib-metadata==4.13.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupPy==1.14
MarkupSafe==2.1.1
msgpack==1.0.4
netaddr==0.7.19
numpy==1.23.4
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
openwisp-users==1.0.2
openwisp-utils==1.0.4
packaging==21.3
pandas==1.5.0
phonenumbers==8.12.56
Pillow==9.2.0
prompt-toolkit==3.0.32
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.5.0
pymongo==3.10.1
mongoengine==0.24.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-ldap==3.2.0
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
reversion==0.2
rjsmin==1.2.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tablib==3.2.1
toml==0.10.2
types-cryptography==3.3.23
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xadmin-py3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
supervisor==4.2.4
daphne==2.5.0